[AIRFLOW-3461] Move TaskFail out of models.py (#4630)
