+++ airflow/api/common/experimental/delete_dag.py
@@ -20,6 +20,7 @@
 from sqlalchemy import or_
 
 from airflow import models
+from airflow.models.taskfail import TaskFail
 from airflow.utils.db import provide_session
 from airflow.exceptions import DagNotFound, DagFileExists
 
@@ -56,7 +57,7 @@ def delete_dag(dag_id, keep_records_in_log=True, session=None):
 
     if dag.is_subdag:
         p, c = dag_id.rsplit(".", 1)
-        for m in models.DagRun, models.TaskFail, models.TaskInstance:
+        for m in models.DagRun, TaskFail, models.TaskInstance:
             count += session.query(m).filter(m.dag_id == p, m.task_id == c).delete()
 
     return count
