+++ tests/contrib/sensors/test_weekday_sensor.py
@@ -24,6 +24,7 @@ from airflow.contrib.sensors.weekday_sensor import DayOfWeekSensor
 from airflow.contrib.utils.weekday import WeekDay
 from airflow.exceptions import AirflowSensorTimeout
 from airflow.models import DagBag
+from airflow.models.taskfail import TaskFail
 from airflow.settings import Session
 from airflow.utils.timezone import datetime
 
@@ -53,7 +54,7 @@ class DayOfWeekSensorTests(unittest.TestCase):
         session = Session()
         session.query(models.TaskInstance).filter_by(
             dag_id=TEST_DAG_ID).delete()
-        session.query(models.TaskFail).filter_by(
+        session.query(TaskFail).filter_by(
             dag_id=TEST_DAG_ID).delete()
         session.commit()
         session.close()
